+++ relations.js
@@ -1,0 +1,3 @@
+const assertRelations = () => {};
+
+export default assertRelations;
